add: [logger] user management API
API の URL を変更 (resolve #3)

+++ logger/api.py
@@ -12,10 +12,15 @@ class Order(ndb.Model):
   content = ndb.JsonProperty()
 
 
+def getUserName():
+  user = users.get_current_user()
+  return user.nickname() if user else 'anonymous'
+
+
 class OrderListHandler(webapp2.RequestHandler):
   def get(self, username):
-    if username != u'anonymous':
-      return self.error(404)
+    if not username:
+      username = getUserName()
     orders = Order.query(Order.username == username).order(-Order.date).iter()
 
     orderList = []
@@ -28,45 +33,55 @@ class OrderListHandler(webapp2.RequestHandler):
 
 
 class OrderHandler(webapp2.RequestHandler):
-  def respond(self, order):
-    content = order.content
+  def update(self, order):
+    content = json.loads(self.request.body)
+    order.date = datetime.strptime(content['date'], '%Y-%m-%d')
+    order.content = content
+    order.put()
     content['id'] = order.key.id()
     self.response.headers['Content-Type'] = 'application/json'
     self.response.out.write(json.dumps(content))
 
-  def get(self, username, orderid):
-    if orderid.isdigit():
+  def post(self, method, orderid):
+    if method == 'create':
+      order = Order(username=getUserName())
+      self.update(order)
+      return
+    elif method == 'update':
       order = Order.get_by_id(int(orderid))
-      if order:
-        self.respond(order)
+      if order and order.username == getUserName():
+        self.update(order)
         return
-    self.error(404)
-
-  def put(self, username, orderid):
-    if orderid.isdigit():
+    elif method == 'delete':
       order = Order.get_by_id(int(orderid))
-    else:
-      order = Order(username=username)
-    content = json.loads(self.request.body)
-    if content['date']:
-      order.date = datetime.strptime(content['date'], '%Y-%m-%d')
-    else:
-      order.date = datetime.now()
-    order.content = content
-    order.put()
-    self.respond(order)
-
-  def delete(self, username, orderid):
-    if orderid.isdigit():
-      order = Order.get_by_id(int(orderid))
-      if order:
+      if order and order.username == getUserName():
         order.key.delete()
         return
     self.error(404)
 
 
+class LogInHandler(webapp2.RequestHandler):
+  def get(self, type):
+    if type == 'in':
+      self.redirect(users.create_login_url('/logger/'))
+    elif type == 'out':
+      self.redirect(users.create_logout_url('/logger/'))
+
+
+class UserHandler(webapp2.RequestHandler):
+  def get(self):
+    username = getUserName()
+    self.response.headers['Content-Type'] = 'application/json'
+    self.response.out.write(json.dumps({
+      'name': username,
+      'isLoggedIn': username != 'anonymous'
+    }))
+
 
 app = webapp2.WSGIApplication([
-  (r'/logger/api/(\w+)', OrderListHandler),
-  (r'/logger/api/(\w+)/(\w+)', OrderHandler)
+  (r'/logger/log(in|out)', LogInHandler),
+  (r'/logger/api/verify', UserHandler),
+  (r'/logger/api/orders/?(\w+)?', OrderListHandler),
+  (r'/logger/api/order/(create)()', OrderHandler),
+  (r'/logger/api/order/(\w+)/(\d+)', OrderHandler)
 ], debug=True)
